requirements.txt
pyTelegramBotAPI==4.21.0
telethon==1.36.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
Pillow==10.4.0
qrcode==7.4.2
loguru==0.7.2
requests==2.32.3
cryptography==43.0.1
bcrypt==4.2.0
passlib==1.7.4
